Modify build scripts for clang. Disable -mips16 flag for now. Fix a few issues and incompatibilities detected by compiler.

+++ include/strings.h
@@ -37,7 +37,7 @@ typedef unsigned size_t;
 
 int      bcmp(const void *, const void *, size_t);
 void     bcopy(const void *, void *, size_t);
-void     bzero(void *, unsigned long);
+void     bzero(void *, unsigned int);
 int      ffs(int);
 char    *index(const char *, int);
 char    *rindex(const char *, int);
